+++ src/main/java/woopaca/jspcafe/service/PostService.java
@@ -3,9 +3,9 @@ package woopaca.jspcafe.service;
 import woopaca.jspcafe.model.Post;
 import woopaca.jspcafe.repository.PostRepository;
 import woopaca.jspcafe.repository.UserRepository;
-import woopaca.jspcafe.servlet.dto.PostDetailsResponse;
-import woopaca.jspcafe.servlet.dto.PostsResponse;
-import woopaca.jspcafe.servlet.dto.WritePostRequest;
+import woopaca.jspcafe.servlet.dto.response.PostDetailsResponse;
+import woopaca.jspcafe.servlet.dto.response.PostsResponse;
+import woopaca.jspcafe.servlet.dto.request.WritePostRequest;
 
 import java.util.List;
