+++ src/main/java/woopaca/jspcafe/servlet/UserProfileServlet.java
@@ -8,7 +8,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import woopaca.jspcafe.service.UserService;
-import woopaca.jspcafe.servlet.dto.UserProfile;
+import woopaca.jspcafe.servlet.dto.response.UserProfile;
 
 import java.io.IOException;
