+++ src/main/java/woopaca/jspcafe/servlet/dto/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package woopaca.jspcafe.servlet.dto;
+package woopaca.jspcafe.servlet.dto.request;
 
 public record SignUpRequest(String username, String nickname, String password) {
 }
