+++ src/main/java/woopaca/jspcafe/servlet/dto/request/WritePostRequest.java
@@ -1,4 +1,4 @@
-package woopaca.jspcafe.servlet.dto;
+package woopaca.jspcafe.servlet.dto.request;
 
 public record WritePostRequest(String title, String content) {
 }
